+++ default.nix
@@ -77,6 +77,7 @@ with pkgs; with python3.pkgs; buildPythonPackage rec {
     tkinter
     lxml
     deprecated
+    immutabledict
   ];
 
   computed_PATH = lib.makeBinPath propagatedBuildInputs;
