+++ openlane/config/__init__.py
@@ -21,4 +21,4 @@ configuration objects. Configuration objects are the primary input to a flow.
 from .preprocessor import Keys
 from .variable import Instance, Macro, Variable
 from .config import Meta, Config, InvalidConfig
-from .flow import all_variables as universal_flow_config_variables
+from .flow import flow_common_variables as universal_flow_config_variables
