+++ openlane/steps/klayout.py
@@ -54,8 +54,101 @@ def get_lef_args(config: Config, toolbox: Toolbox) -> List[str]:
     return lef_args
 
 
+class KLayoutStep(Step):
+    config_vars = [
+        Variable(
+            "KLAYOUT_TECH",
+            Optional[Path],
+            "A path to the KLayout layer technology (.lyt) file.",
+            pdk=True,
+        ),
+        Variable(
+            "KLAYOUT_PROPERTIES",
+            Optional[Path],
+            "A path to the KLayout layer properties (.lyp) file.",
+            pdk=True,
+        ),
+        Variable(
+            "KLAYOUT_DRC_TECH_SCRIPT",
+            Optional[Path],
+            "A path to a KLayout DRC tech script.",
+            pdk=True,
+        ),
+        Variable(
+            "KLAYOUT_DEF_LAYER_MAP",
+            Optional[Path],
+            "A path to the KLayout LEF/DEF layer mapping (.map) file.",
+            pdk=True,
+        ),
+        Variable(
+            "KLAYOUT_XOR_IGNORE_LAYERS",
+            Optional[List[str]],
+            "KLayout layers to ignore during XOR operations.",
+            pdk=True,
+        ),
+    ]
+
+
 @Step.factory.register()
-class StreamOut(Step):
+class Render(KLayoutStep):
+    """
+    Renders a PNG of the layout using KLayout.
+
+    DEF is required as an input, but if a GDS-II view
+    exists in the input state, it will be used instead.
+    """
+
+    id = "KLayout.Render"
+    name = "Render Image (w/ KLayout)"
+
+    inputs = [DesignFormat.DEF]
+    outputs = []
+
+    def run(self, state_in: State, **kwargs) -> Tuple[ViewsUpdate, MetricsUpdate]:
+        input_view = state_in[DesignFormat.DEF]
+        if gds := state_in[DesignFormat.GDS]:
+            input_view = gds
+
+        lyp = self.config["KLAYOUT_PROPERTIES"]
+        lyt = self.config["KLAYOUT_TECH"]
+        lym = self.config["KLAYOUT_DEF_LAYER_MAP"]
+
+        tech_lefs = self.toolbox.filter_views(self.config, self.config["TECH_LEFS"])
+        if len(tech_lefs) != 1:
+            raise StepError(
+                "Misconfigured SCL: 'TECH_LEFS' must return exactly one Tech LEF for its default timing corner."
+            )
+
+        lef_arguments = ["-l", str(tech_lefs[0])]
+        for file in self.config["CELL_LEFS"]:
+            lef_arguments += ["-l", str(file)]
+        if extra := self.config["EXTRA_LEFS"]:
+            for file in extra:
+                lef_arguments += ["-l", str(file)]
+
+        self.run_subprocess(
+            [
+                sys.executable,
+                os.path.join(get_script_dir(), "klayout", "render.py"),
+                input_view,
+                "--output",
+                os.path.join(self.step_dir, "out.png"),
+                "--lyp",
+                lyp,
+                "--lyt",
+                lyt,
+                "--lym",
+                lym,
+            ]
+            + lef_arguments,
+            silent=True,
+        )
+
+        return {}, {}
+
+
+@Step.factory.register()
+class StreamOut(KLayoutStep):
     """
     Converts DEF views into GDSII streams using KLayout.
 
@@ -74,7 +167,7 @@ class StreamOut(Step):
     inputs = [DesignFormat.DEF]
     outputs = [DesignFormat.GDS, DesignFormat.KLAYOUT_GDS]
 
-    config_vars = [
+    config_vars = KLayoutStep.config_vars + [
         Variable(
             "RUN_KLAYOUT_STREAMOUT",
             bool,
@@ -157,9 +250,7 @@ class StreamOut(Step):
             return None
         assert self.toolbox is not None
 
-        if image := self.toolbox.render_png(
-            self.config, str(self.state_out["klayout_gds"])
-        ):
+        if image := self.toolbox.render_png(self.config, self.state_out):
             image_encoded = b64encode(image).decode("utf8")
             return f'<img src="data:image/png;base64,{image_encoded}" />'
 
@@ -167,7 +258,7 @@ class StreamOut(Step):
 
 
 @Step.factory.register()
-class XOR(Step):
+class XOR(KLayoutStep):
     """
     Performs an XOR operation on the Magic and KLayout GDS views. The idea is:
     if there's any difference between the GDSII streams between the two tools,
@@ -184,7 +275,7 @@ class XOR(Step):
     ]
     outputs = []
 
-    config_vars = [
+    config_vars = KLayoutStep.config_vars + [
         Variable(
             "RUN_KLAYOUT_XOR",
             bool,
@@ -243,7 +334,7 @@ class XOR(Step):
 
 
 @Step.factory.register()
-class OpenGUI(Step):
+class OpenGUI(KLayoutStep):
     """
     Opens the DEF view in the KLayout GUI, with layers loaded and mapped
     properly. Useful to inspect ``.klayout.xml`` database files and the like.
